refactor: #671 컴포즈로 마이그레이션된 xml 제거 및 패키지 이동
