+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/discussion/hot/HotDiscussionScreen.kt
@@ -16,7 +16,7 @@ import com.team.todoktodok.presentation.compose.component.InfinityLazyColumn
 import com.team.todoktodok.presentation.compose.discussion.activate.ActivatedDiscussionHeader
 import com.team.todoktodok.presentation.compose.discussion.popular.PopularDiscussionHeader
 import com.team.todoktodok.presentation.compose.preview.HotDiscussionPreviewParameterProvider
-import com.team.todoktodok.presentation.xml.discussions.hot.HotDiscussionUiState
+import com.team.todoktodok.presentation.compose.discussion.hot.HotDiscussionUiState
 
 @Composable
 fun HotDiscussionScreen(
