+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/presentation/compose/preview/HotDiscussionPreviewParameterProvider.kt
@@ -1,7 +1,7 @@
 package com.team.todoktodok.presentation.compose.preview
 
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import com.team.todoktodok.presentation.xml.discussions.hot.HotDiscussionUiState
+import com.team.todoktodok.presentation.compose.discussion.hot.HotDiscussionUiState
 
 class HotDiscussionPreviewParameterProvider : PreviewParameterProvider<HotDiscussionUiState> {
     override val values: Sequence<HotDiscussionUiState> =
